fastapi/requirements.txt
numpy
fastapi==0.88.0
scikit-learn
uvicorn
kafka-python